benchmark

+++ SLIDE-x-BENCH/KERNEL/banker_algorithm/frst.c
@@ -1,8 +1,8 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef int64_t TARGET_TYPE;
-typedef uint64_t TARGET_INDEX;
+typedef double TARGET_TYPE;
+typedef uint8_t TARGET_INDEX;
 
 
 TARGET_TYPE need[size][size];
